+++ packages/headless/src/features/triggers/triggers-slice.ts
@@ -1,21 +1,22 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {executeSearch} from '../search/search-actions';
 import {getTriggerInitialState} from './triggers-state';
-import {
-  Trigger,
-  TriggerRedirect,
-  isTriggerRedirect,
-} from './../../api/search/trigger';
+import {isRedirectTrigger, isQueryTrigger} from './../../api/search/trigger';
 
 export const triggerReducer = createReducer(
   getTriggerInitialState(),
   (builder) =>
     builder.addCase(executeSearch.fulfilled, (state, action) => {
-      const redirectTriggers: Trigger[] = action.payload.response.triggers.filter(
-        (trigger) => isTriggerRedirect(trigger)
+      const redirectTriggers = action.payload.response.triggers.filter(
+        isRedirectTrigger
       );
       state.redirectTo = redirectTriggers.length
-        ? (redirectTriggers[0] as TriggerRedirect).content
+        ? redirectTriggers[0].content
         : '';
+
+      const queryTriggers = action.payload.response.triggers.filter(
+        isQueryTrigger
+      );
+      state.query = queryTriggers.length ? queryTriggers[0].content : '';
     })
 );