+++ packages/headless/src/features/triggers/triggers-state.ts
@@ -2,9 +2,15 @@ export interface TriggerState {
   /**
    * The URL to redirect the user to after receiving a redirection trigger.
    */
-  redirectTo: string | null;
+  redirectTo: string;
+
+  /**
+   * The new query to perform a search with after receiving a query trigger.
+   */
+  query: string;
 }
 
 export const getTriggerInitialState: () => TriggerState = () => ({
-  redirectTo: null,
+  redirectTo: '',
+  query: '',
 });